[RW-2355][risk=no] Convert detail-tabs component to react, match new design (#2114)
* RW-2355 convert detail-tabs component to react

* RW-2355 override reuse condition for participant detail routes

* RW-2355 style tabs and content to match new design

* RW-2355 delete unused files

* RW-2355 remove angular references in detail table and chart components

* RW-2355 fix broken test

* [RW-2627][Risk=low] Allow access to data use agreement when not registered (#2117)

* [RW-2622][risk=low] Add consume project to Billing Project Buffer API (#2115)

* add code and test for the most basic functionality. no locking and error handling

* add locks on assignment and test

* clean up no transaction test

* codacy

* code review - cleanup test annotation and synchronized test locks

* code review - final modifier and comment

* codacy

* [RW-2619][Risk=low] Remove workspace storage service. (#2112)

* [RW-2619][Risk=low] Remove workspace storage service.

* Fix typing errors in more tests

* Fix one more thing

* PR feedback

* Fix tests

* Working tests

* More PR feedback

* Change to promise

* [No Ticket][risk=no] Replace deprecated Mockito Matchers with ArgumentMatchers (#2086)

* [risk=no] adding type and subtype to dao call. (#2120)

* adding type and subtype to dao call.

* [RW-2619][Risk=low] Remove workspace storage service. (#2112)

* [RW-2619][Risk=low] Remove workspace storage service.

* Fix typing errors in more tests

* Fix one more thing

* PR feedback

* Fix tests

* Working tests

* More PR feedback

* Change to promise

* [No Ticket][risk=no] Replace deprecated Mockito Matchers with ArgumentMatchers (#2086)

* fixed test.

* adding type and subtype to dao call.

* fixed test.

* [risk=no] adding new dao calls for criteria. (#2119)

* adding new dao calls for criteria.

* removed unused import.

* [RW-2619][Risk=low] Remove workspace storage service. (#2112)

* [RW-2619][Risk=low] Remove workspace storage service.

* Fix typing errors in more tests

* Fix one more thing

* PR feedback

* Fix tests

* Working tests

* More PR feedback

* Change to promise

* [No Ticket][risk=no] Replace deprecated Mockito Matchers with ArgumentMatchers (#2086)

* adding new dao calls for criteria.

* removed unused import.

* trying to get codacy to rerun.

* [risk=no] adding type and subtype to dao call. (#2120)

* adding type and subtype to dao call.

* [RW-2619][Risk=low] Remove workspace storage service. (#2112)

* [RW-2619][Risk=low] Remove workspace storage service.

* Fix typing errors in more tests

* Fix one more thing

* PR feedback

* Fix tests

* Working tests

* More PR feedback

* Change to promise

* [No Ticket][risk=no] Replace deprecated Mockito Matchers with ArgumentMatchers (#2086)

* fixed test.

* adding type and subtype to dao call.

* fixed test.

* adding new dao calls for criteria.

* adding new dao calls for criteria.

* removed unused import.

* removed unused import.

* trying to get codacy to rerun.

* [risk=no] add new indexes in cb_criteria (#2121)

* adding new indexes.

* adding new indexes.

* [risk=no] update search parameter (#2124)

* updating search parameter

* updating search parameter

* [RW-2392] [risk=no] Improve stability of access module completion timestamps (#2080)

* Improve the logging of UserService data access level changes and access module sync methods.
* Add improved logging to the offline user cron tasks.
* Fix the eRA Commons and Moodle sync methods to avoid updating completion time on every check.
* Add tests for completion timestamp stability.
* Add a note documenting the NihStatus expire time.
* Slightly adjust the compliance timestamp update logic, & fix unit tests.

* Format value for Timestamp (#2116)

* [RW-2622][risk=low]  add forgotten liquibase migration from consume buffer PR (#2126)

* RW-2355 convert detail-tabs component to react

* RW-2355 override reuse condition for participant detail routes

* RW-2355 style tabs and content to match new design

* RW-2355 delete unused files

* RW-2355 remove angular references in detail table and chart components

* RW-2355 fix broken test

* RW-2355 fix import
