+++ ui/src/app/cohort-review/cohort-review.module.ts
@@ -18,9 +18,7 @@ import {CreateReviewPage} from './create-review-page/create-review-page';
 // This is a temporary measure until we have specs and APIs for overview specific charts
 import {DetailHeaderComponent} from './detail-header/detail-header.component';
 import {DetailPage} from './detail-page/detail-page';
-import {DetailTabTableComponent} from './detail-tab-table/detail-tab-table.component';
 import {DetailTabsComponent} from './detail-tabs/detail-tabs.component';
-import {IndividualParticipantsChartsComponent} from './individual-participants-charts/individual-participants-charts';
 import {PageLayout} from './page-layout/page-layout';
 import {CohortReviewRoutingModule} from './routing/routing.module';
 import {SidebarContentComponent} from './sidebar-content/sidebar-content.component';
@@ -58,12 +56,10 @@ import {TablePage} from './table-page/table-page';
     EditAnnotationDefinitionsModalComponent,
 
     /* Participant Table */
-    IndividualParticipantsChartsComponent,
     /* Participant Detail */
     SidebarContentComponent,
     DetailHeaderComponent,
     DetailTabsComponent,
-    DetailTabTableComponent,
   ],
   providers: [
     {
