+++ ui/src/app/cohort-review/detail-tab-table/detail-tab-table.component.tsx
@@ -1,11 +1,10 @@
-import {Component, Input} from '@angular/core';
 import {ReviewDomainChartsComponent} from 'app/cohort-review/review-domain-charts/review-domain-charts';
 import {vocabOptions} from 'app/cohort-review/review-state.service';
 import {datatableStyles} from 'app/cohort-review/review-utils/primeReactCss.utils';
 import {TextInput} from 'app/components/inputs';
 import {SpinnerOverlay} from 'app/components/spinners';
 import {cohortReviewApi} from 'app/services/swagger-fetch-clients';
-import {reactStyles, ReactWrapperBase, withCurrentWorkspace} from 'app/utils';
+import {reactStyles, withCurrentWorkspace} from 'app/utils';
 import {urlParamsStore} from 'app/utils/navigation';
 import {WorkspaceData} from 'app/utils/workspace-data';
 import {DomainType, PageFilterRequest, PageFilterType, SortOrder} from 'generated/fetch';
@@ -164,6 +163,7 @@ const filterIcons = {
     color: '#262262',
   }
 };
+
 const rows = 25;
 const domains = [
   DomainType.CONDITION,
@@ -702,31 +702,3 @@ export const DetailTabTable = withCurrentWorkspace()(
     }
   }
 );
-
-@Component ({
-  selector : 'app-detail-tab-table',
-  template: '<div #root></div>'
-})
-export class DetailTabTableComponent extends ReactWrapperBase {
-  @Input('tabName') tabName: DetailTabTableProps['tabName'];
-  @Input('columns') columns: DetailTabTableProps['columns'];
-  @Input('domain') domain: DetailTabTableProps['domain'];
-  @Input('filterType') filterType: DetailTabTableProps['filterType'];
-  @Input('participantId') participantId: DetailTabTableProps['participantId'];
-  @Input('filterState') filterState: DetailTabTableProps['filterState'];
-  @Input('getFilteredData') getFilteredData: DetailTabTableProps['getFilteredData'];
-  @Input('updateState') updateState: DetailTabTableProps['updateState'];
-
-  constructor() {
-    super(DetailTabTable, [
-      'tabName',
-      'columns',
-      'domain',
-      'filterType',
-      'participantId',
-      'filterState',
-      'getFilteredData',
-      'updateState',
-    ]);
-  }
-}
