--- ui/src/app/cohort-review/detail-tabs/detail-tabs.component.css
@@ -1,13 +1,0 @@
-
-.details {
-  border: 1px solid gray;
-  margin: 1rem 0 0 1rem;
-  background-color: white;
-  padding: 0.2rem;
-  min-width: 12rem;
-}
-
-.summary-container {
-  margin: 2rem 0rem;
-  min-width: 100%;
-}
