--- ui/src/app/cohort-review/detail-tabs/detail-tabs.component.html
@@ -1,46 +1,0 @@
-<clr-tabs>
-  <clr-tab>
-    <button clrTabLink id="detail-summary-tab">Summary</button>
-    <ng-template [(clrIfActive)]="summaryActive">
-      <clr-tab-content id="detail-summary-content" *clrIfActive>
-        <div class="container summary-container">
-          <div class="row">
-            <div class="col-sm-4" *ngFor = "let charts of domainList; let i=index">
-              <div *ngIf="chartData[charts].loading" class="loading-spinner spinner spinner-lg spin-in-overlay">
-                Loading...
-              </div>
-              <clr-alert *ngIf="!chartData[charts].loading && !chartData[charts].items.length" [clrAlertClosable]="false">
-                <clr-alert-item>
-                  <span class="alert-text">
-                    There are no {{chartData[charts].conditionTitle}} to show for this participant.
-                  </span>
-                </clr-alert-item>
-              </clr-alert>
-              <app-individual-participants-charts [chartData]="chartData[charts]">
-              </app-individual-participants-charts>
-            </div>
-          </div>
-        </div>
-      </clr-tab-content>
-    </ng-template>
-  </clr-tab>
-
-  <!-- Individual Tab Views -->
-  <clr-tab *ngFor="let tab of tabs">
-    <button clrTabLink id="detail-{{ tab.name | lowercase }}-tab" title="{{ tab.name | titlecase }}">
-      {{ tab.name | titlecase }}
-    </button>
-    <clr-tab-content id="detail-{{ tab.name | lowercase }}-content" *clrIfActive>
-      <app-detail-tab-table
-        [getFilteredData]=filteredData
-        [filterState]="filterState"
-        [updateState]="updateState"
-        [tabName]="tab.name"
-        [columns]="tab.columns[vocab]"
-        [filterType]="tab.filterType"
-        [domain]="tab.domain"
-        [participantId]="participantId">
-      </app-detail-tab-table>
-    </clr-tab-content>
-  </clr-tab>
-</clr-tabs>
