--- ui/src/app/cohort-review/individual-participants-charts/individual-participants-charts.css
@@ -1,4 +1,0 @@
-.chart-container {
-  display:block;
-  width: 90% !important;
-}
