+++ ui/src/app/cohort-review/review-utils/primeReactCss.utils.ts
@@ -25,6 +25,7 @@ export const datatableStyles = `
   }
   .p-datatable .p-datatable-scrollable-wrapper {
     border: 1px solid #c8c8c8;
+    border-radius: 3px;
   }
   .p-datatable .p-paginator.p-paginator-bottom {
     border: 0;
